mavlink_core

Enum ConnectionAddress

pub enum ConnectionAddress {
    Tcp(TcpConfig),
    Udp(UdpConfig),
    Serial(SerialConfig),
    File(FileConfig),
}

Available on crate features std or tokio-1 only.

A parsed MAVLink connection address

Variants

Tcp(TcpConfig)

Available on crate feature tcp only.

TCP client or server address

Udp(UdpConfig)

Available on crate feature udp only.

UDP client, server or broadcast address

Serial(SerialConfig)

Available on crate feature direct-serial only.

Serial port address

File(FileConfig)

File input address

Implementations

impl ConnectionAddress

pub fn parse_address(address: &str) -> Result<Self, Error>

Parse a MAVLink address string.

The address must be in one of the following formats:

tcpin:<addr>:<port> to create a TCP server, listening for an incoming connection

tcpout:<addr>:<port> to create a TCP client

udpin:<addr>:<port> to create a UDP server, listening for incoming packets

udpout:<addr>:<port> to create a UDP client

udpbcast:<addr>:<port> to create a UDP broadcast

serial:<port>:<baudrate> to create a serial connection

file:<path> to extract file data, writing to such a connection does nothing

Errors

AddrNotAvailable if the address string could not be parsed as a valid MAVLink address