mavlink_core

Module peek_reader

This module implements a buffered/peekable reader.

The purpose of the buffered/peekable reader is to allow for backtracking parsers.

A reader implementing the standard library’s std::io::BufRead trait seems like a good fit, but it does not allow for peeking a specific number of bytes, so it provides no way to request more data from the underlying reader without consuming the existing data.

This API still tries to adhere to the std::io::BufRead’s trait philosophy.

The main type PeekReaderdoes not implement std::io::Read itself, as there is no added benefit in doing so.

Structs

PeekReader

A buffered/peekable reader