mavlink_core

Trait MavConnection

pub trait MavConnection<M: Message> {
    // Required methods
    fn recv(&self) -> Result<(MavHeader, M), MessageReadError>;
    fn recv_raw(&self) -> Result<MAVLinkMessageRaw, MessageReadError>;
    fn try_recv(&self) -> Result<(MavHeader, M), MessageReadError>;
    fn send(
        &self,
        header: &MavHeader,
        data: &M,
    ) -> Result<usize, MessageWriteError>;
    fn set_protocol_version(&mut self, version: MavlinkVersion);
    fn protocol_version(&self) -> MavlinkVersion;
    fn set_allow_recv_any_version(&mut self, allow: bool);
    fn allow_recv_any_version(&self) -> bool;
    fn setup_signing(&mut self, signing_data: Option<SigningConfig>);

    // Provided methods
    fn send_frame(
        &self,
        frame: &MavFrame<M>,
    ) -> Result<usize, MessageWriteError> { ... }
    fn recv_frame(&self) -> Result<MavFrame<M>, MessageReadError> { ... }
    fn send_default(&self, data: &M) -> Result<usize, MessageWriteError> { ... }
}

Available on crate feature std only.

A MAVLink connection

Required Methods

fn recv(&self) -> Result<(MavHeader, M), MessageReadError>

Receive a MAVLink message.

May blocks until a valid frame is received, ignoring invalid messages.

Errors

If the connection type blocks until a valid message is received this can not return any errors, otherwise return any errors that occured while receiving.

fn recv_raw(&self) -> Result<MAVLinkMessageRaw, MessageReadError>

Receive a raw, unparsed MAVLink message.

Blocks until a valid frame is received, ignoring invalid messages.

Errors

If the connection type blocks until a valid message is received this can not return any errors, otherwise return any errors that occured while receiving.

fn try_recv(&self) -> Result<(MavHeader, M), MessageReadError>

Try to receive a MAVLink message.

Non-blocking variant of recv(), returns immediately with a MessageReadError if there is an error or no message is available.

Errors

Returns any eror encounter while receiving or deserializing a message

fn send(&self, header: &MavHeader, data: &M) -> Result<usize, MessageWriteError>

Send a MAVLink message

Errors

This function will return a MessageWriteError::Io error when sending fails.

fn set_protocol_version(&mut self, version: MavlinkVersion)

Sets the MAVLink version to use for receiving (when allow_recv_any_version() is false) and sending messages.

fn protocol_version(&self) -> MavlinkVersion

Gets the currently used MAVLink version

fn set_allow_recv_any_version(&mut self, allow: bool)

Set wether MAVLink messages of either version may be received.

If set to false only messages of the version configured with set_protocol_version() are received.

fn allow_recv_any_version(&self) -> bool

Wether messages of any MAVLink version may be received

fn setup_signing(&mut self, signing_data: Option<SigningConfig>)

Available on crate feature signing only.

Setup secret key used for message signing, or disable message signing

Provided Methods

fn send_frame(&self, frame: &MavFrame<M>) -> Result<usize, MessageWriteError>

Write whole frame

Errors

This function will return a MessageWriteError::Io error when sending fails.

fn recv_frame(&self) -> Result<MavFrame<M>, MessageReadError>

Read whole frame

Errors

Returns any eror encounter while receiving or deserializing a message

fn send_default(&self, data: &M) -> Result<usize, MessageWriteError>

Send a message with default header

Errors

This function will return a MessageWriteError::Io error when sending fails.