mavlink_core::types

Struct CharArray

pub struct CharArray<const N: usize> { /* private fields */ }

Abstraction around a byte array that represents a string.

MAVLink encodes strings as C char arrays and the handling is field dependent. This abstration allows to choose if one wants to handle the field as a raw byte array or if one wants the convenience of a str that stops at the first null byte.

Example

use mavlink_core::types::CharArray;

let data = [0x48, 0x45, 0x4c, 0x4c, 0x4f, 0x00, 0x57, 0x4f, 0x52, 0x4c, 0x44, 0x00, 0x00, 0x00];
let ca = CharArray::new(data);
assert_eq!(ca.to_str().unwrap(), "HELLO");
// or using the from str method
let ca: CharArray<10> = "HELLO!".into();
assert_eq!(ca.to_str().unwrap(), "HELLO!");

Implementations

impl<const N: usize> CharArray<N>

pub const fn new(data: [u8; N]) -> Self

pub fn to_str(&self) -> Result<&str, Utf8Error>

Get the string representation of the char array. Returns the string stopping at the first null byte and if the string is not valid utf8 the returned string will be empty.