Module mavlink_core::utils

Traits

RustDefault

A trait very similar to Default but is only implemented for the equivalent Rust types to MavTypes. This is only needed because rust doesn’t currently implement Default for arrays of all sizes. In particular this trait is only ever used when the “serde” feature is enabled. For more information, check out this issue.

Functions

remove_trailing_zeroes

Removes the trailing zeroes in the payload