Trait mavlink_core::utils::RustDefault

pub trait RustDefault: Copy {
    // Required method
    fn rust_default() -> Self;
}

A trait very similar to Default but is only implemented for the equivalent Rust types to MavTypes. This is only needed because rust doesn’t currently implement Default for arrays of all sizes. In particular this trait is only ever used when the “serde” feature is enabled. For more information, check out this issue.

Required Methods

fn rust_default() -> Self

Object Safety

This trait is not object safe.