mavlink_core

Function connect

pub fn connect<M: Message + Sync + Send>(
    address: &str,
) -> Result<Box<dyn MavConnection<M> + Sync + Send>>

Available on crate feature std only.

Connect to a MAVLink node by address string.

The address must be in one of the following formats:

tcpin:<addr>:<port> to create a TCP server, listening an incoming connection

tcpout:<addr>:<port> to create a TCP client

udpin:<addr>:<port> to create a UDP server, listening for incoming packets

udpout:<addr>:<port> to create a UDP client

udpbcast:<addr>:<port> to create a UDP broadcast

serial:<port>:<baudrate> to create a serial connection

file:<path> to extract file data, writing to such a connection does nothing

The type of the connection is determined at runtime based on the address type, so the connection is returned as a trait object.