mavlink_core::utils

Trait RustDefault

pub trait RustDefault: Copy {
    // Required method
    fn rust_default() -> Self;
}

A trait very similar to Default but is only implemented for the equivalent Rust types to MavTypes.

This is only needed because rust doesn’t currently implement Default for arrays of all sizes. In particular this trait is only ever used when the “serde” feature is enabled. For more information, check out this issue.

Required Methods

fn rust_default() -> Self

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.